Ensure ServerProperties default values does not override
Since ServerProperties had primitive properties for port (in
particular) it was not possible to check when applying those
properties if the user had actually changed the value. This
in turn meant that a custom EmbeddedServletContainerFactory
could not set the default values.

Fixed by making int properties of ServerProperties into
Integer and checking for null before setting on the
container factory.

Fixes gh-84

+++ spring-boot/src/test/java/org/springframework/boot/context/embedded/properties/ServerPropertiesTests.java
@@ -24,7 +24,6 @@ import java.util.Map;
 import org.junit.Test;
 import org.springframework.beans.MutablePropertyValues;
 import org.springframework.boot.bind.RelaxedDataBinder;
-import org.springframework.boot.context.embedded.properties.ServerProperties;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -51,7 +50,7 @@ public class ServerPropertiesTests {
 	public void testPortBinding() throws Exception {
 		new RelaxedDataBinder(this.properties, "server").bind(new MutablePropertyValues(
 				Collections.singletonMap("server.port", "9000")));
-		assertEquals(9000, this.properties.getPort());
+		assertEquals(9000, this.properties.getPort().intValue());
 	}
 
 	@Test
